added float 32 cast

+++ sharktank/tests/models/llama/test_llama.py
@@ -46,6 +46,6 @@ def test_llama():
     logits = logits[:, :seq_len, :]
 
     ids = ids[0, 1:]
-    logits = logits[0, :-1]
+    logits = logits[0, :-1].to(torch.float32)
     cross_entropy = torch.nn.functional.cross_entropy(logits, ids)
     assert pytest.approx(0.577, 1e-2) == cross_entropy
